+++ CHANGELOG.md
@@ -12,3 +12,17 @@
 
 - 📄 Use raw image in *README*.
 - ⚙️ Extract default value to *default.ts*, and destruct *PI* and *round()* from *Math*.
+
+## 0.2.1
+
+`2020-10-10`
+
+- 🆕 Support dynamic percentage animation.
+
+## 0.3.0
+
+`2020-10-15`
+
+- 🆕 Add `onAnimationStart` and `onAnimationEnd` lifecycles.
+- ⚙️ Modify default data value.
+- 💿 Perfect examples.